fixed Issue 311:  	 background colors ignored/overwritten from/to config file Also I added BodyGetOpacity and BodySetOpacity. Also, I tidied up some code by removing HeeksColor::glMaterial, because this is a duplicate of constructing a "Material" from a HeeksColor and using its' glMaterial function.

+++ trunk/src/HCircle.cpp
@@ -286,7 +286,7 @@ HeeksObj* HCircle::ReadFromXMLElement(TiXmlElement* pElem)
 	for(TiXmlAttribute* a = pElem->FirstAttribute(); a; a = a->Next())
 	{
 		std::string name(a->Name());
-		if(name == "col"){c = HeeksColor(a->IntValue());}
+		if(name == "col"){c = HeeksColor((long)(a->IntValue()));}
 		else if(name == "r"){r = a->DoubleValue();}
 		else if(name == "cx"){centre.SetX(a->DoubleValue());}
 		else if(name == "cy"){centre.SetY(a->DoubleValue());}
