+++ trunk/src/HEllipse.cpp
@@ -413,7 +413,7 @@ HeeksObj* HEllipse::ReadFromXMLElement(TiXmlElement* pElem)
 	for(TiXmlAttribute* a = pElem->FirstAttribute(); a; a = a->Next())
 	{
 		std::string name(a->Name());
-		if(name == "col"){c = HeeksColor(a->IntValue());}
+		if(name == "col"){c = HeeksColor((long)(a->IntValue()));}
 		else if(name == "maj"){maj = a->DoubleValue();}
 		else if(name == "min"){min = a->DoubleValue();}
 		else if(name == "rot"){rot = a->DoubleValue();}
